+++ lib/sites.py
@@ -8,6 +8,6 @@ def next_random_siteid(num_sites=10000):
   num_domains = len(popular_domains)
   return "http://{0}/page{1}".format(
     popular_domains[random.randint(0, num_domains-1)],
-    random.randint(1, num_sites / num_domains)
+    random.randint(1, int(num_sites / num_domains))
   )
